refactor ToggleableMMI
