+++ Content.Server/_DV/Silicons/MMI/ToggleableMMIComponent.cs
@@ -1,0 +1,9 @@
+using Robust.Shared.GameObjects;
+using Robust.Shared.Serialization;
+
+namespace Content.Server._DV.Silicons.MMI;
+
+[RegisterComponent]
+public sealed partial class ToggleableMMIComponent : Component
+{
+}
